refactor destroy_clones to separate class
